refactor: remove legacy notification code
