+++ src/NotificationProvider.tsx
@@ -1,5 +1,4 @@
-import type { FC } from 'react';
-import React from 'react';
+import * as React from 'react';
 
 export interface NotificationContextProps {
   classNames?: {
@@ -14,10 +13,10 @@ export interface NotificationProviderProps extends NotificationContextProps {
   children: React.ReactNode;
 }
 
-const NotificationProvider: FC<NotificationProviderProps> = ({ children, classNames }) => {
-  return (
-    <NotificationContext.Provider value={{ classNames }}>{children}</NotificationContext.Provider>
-  );
+const NotificationProvider: React.FC<NotificationProviderProps> = ({ children, classNames }) => {
+  const context = React.useMemo<NotificationContextProps>(() => ({ classNames }), [classNames]);
+
+  return <NotificationContext.Provider value={context}>{children}</NotificationContext.Provider>;
 };
 
 export default NotificationProvider;
